feat: Add migration file for school_years table.
Make this addition to create a table in the
database. The table will be called
"school_years". This table will hold all the data
for the different ROTCU school years that will be
added and to keep track of all the information
related to each school year.

The table will contain an id field as the primary
key and it's respective timestamp data.

It will have an integer field to hold the data of
the year that the school year will start at called
'year_start.'

It will have another integer field that is
nullable to hold the year the school year will
end at. It is made nullable to cover the very low
chance that a school year might last more than a
regular calendar year. The field will be
called 'year_end.'

Finally, it will have a nullable text field called
'description'. This field will hold any comments
that the creator may have about the school year.

These additions will store all the necessary data
about a school year.